Add pentax support
Former-commit-id: 9ca290c7cb11f79c23a2bccb47b80f4a254476cf [formerly 705d8e5f426c27c5f0f6ac4b7a4636561a5a548c]
Former-commit-id: a3e70916f95ad9560965267c195bba18b1b017d3 [formerly 293e176c3051ac408a55a9e7b809449e86e16fd4]
Former-commit-id: fffd574a2722595faca31af84065675ec346bfe8
Former-commit-id: 21040908c9b3580faf4859fb682255e3fb2e8c73

+++ Source/Raw.Net/Source/Format/Tiff/IFD/Tag.cs
@@ -128,7 +128,7 @@ namespace RawNet
             dataOffset = 0;
             if (((dataCount * GetTypeSize(dataType) > 4)))
             {
-                dataOffset = (uint)(fileStream.ReadUInt32());
+                dataOffset = (fileStream.ReadUInt32());
             }
             //Get the tag data
             data = new Object[dataCount];
